kafka-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.33.0
azure-storage-blob==12.25.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
isodate==0.7.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
python-multipart==0.0.5
kafka-python==2.1.5
